+++ src/main/java/domain/Operator.java
@@ -1,4 +1,4 @@
-package study;
+package domain;
 
 import java.util.Arrays;
 import java.util.function.BiFunction;
