+++ src/main/java/study/Application.java
@@ -2,6 +2,8 @@ package study;
 
 import java.util.Scanner;
 
+import domain.Calculator;
+
 public class Application {
 	public static void main(String[] args) {
 		Scanner scanner = new Scanner(System.in);
